% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomsky-package.R
\docType{package}
\name{bloomsky}
\alias{bloomsky}
\alias{bloomsky-package}
\title{Tools to Work with 'BloomSky' Weather Station Data}
\description{
The 'BloomSky' weather station has a cloud service \url{https://www.bloomsky.com/} that centralizes data
collection and provides a user interface for interactive data viewing as well as an 'API'. Tools are
provided to work with the underlying weather data feeds from the site.
}
\author{
Bob Rudis (bob@rud.is)
}
