% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{bsky_get_reading}
\alias{bsky_get_reading}
\title{Retrieve last, current station reading}
\usage{
bsky_get_reading(bsky_api_key = Sys.getenv("BLOOMSKY_API_KEY"))
}
\arguments{
\item{bsky_api_key}{the API key for BloomSky}
}
\description{
Retrieve last, current station reading
}
\note{
You need \code{BLOOMSKY_API_KEY} set in your R environment. You can retrieve that by
going to \url{http://dashboard.bloomsky.com/}
and copying the value from "Your API Key" into a value in your local \code{.Renviron}.
}
\examples{
bsky_get_reading()
}
